Refactor incorrect of use of helper "modules"
