--- app/lib/llm_clients/openai/completion_request_helper.rb
@@ -1,21 +1,0 @@
-require "openai"
-
-module LlmClients
-  module Openai
-    module CompletionRequestHelper
-      def complete_request(prompt, &)
-        messages = []
-        messages << { role: "user", content: prompt }
-
-        response = chat_client.chat(
-          parameters: {
-            model: @model,
-            messages:,
-            temperature: @temperature,
-          })
-        yield response.dig("choices", 0, "message", "content")
-        response
-      end
-    end
-  end
-end
